// src/modules/v1/theory-notebook/enums/theoryNotebook.enum.ts
export enum UnderstandLevelFilterEnum {
  DEFAULT = "0",
  KNOWN = "1",
  UNKNOWN = "2",
  UNSURE = "3",
}

export enum TickEnum {
  TRUE = "1",
  FALSE = "0",
}

export enum TheoryLevelEnum {
  LEVEL_1 = "1",
  LEVEL_2 = "2",
  LEVEL_3 = "3",
  LEVEL_4 = "4",
  LEVEL_5 = "5",
  LEVEL_6 = "6",
  LEVEL_7 = "7",
}

export enum KindFilterEnum {
  HANZII = "hanzii",
  WORD = "word",
  GRAMMAR = "grammar",
}